+++ src/p4est_mesh.h
@@ -107,9 +107,8 @@ void                p4est_order_local_vertices (p4est_t * p4est,
  *
  */
 void                p4est_possible_node_neigbors (p4est_quadrant_t * q,
-                                                  int32_t node,
-                                                  int8_t nnum,
-                                                  int8_t neighbor_rlev,
+                                                  int node, int nnum,
+                                                  int neighbor_rlev,
                                                   p4est_quadrant_t * neighbor,
                                                   int *neighbor_node);
